Add TensorFlow test file (WIP)
I followed TensorFlow's first example of image classification in
https://www.tensorflow.org/tutorials/keras/basic_classification,
and attempted to figure out how to save the training model so I could export it
to another format. Failing that, I added a lot of comments and links in the code
that I could use to pick this up later.